Split complex functions from API_Client into API_Facade
