+++ src/php/admin/tinymce.php
@@ -103,13 +103,13 @@ function ajax_handler_body() {
  */
 function list_directories_in_path( array $path, $root ) {
 	if ( 0 === count( $path ) ) {
-		return \Sgdg\API_Client::list_directories( $root, new \Sgdg\Frontend\API_Fields( array( 'name' ) ) )->then(
+		return \Sgdg\API_Facade::list_directories( $root, new \Sgdg\Frontend\API_Fields( array( 'name' ) ) )->then(
 			static function( $directories ) {
 				return array_column( $directories, 'name' );
 			}
 		);
 	}
-	return \Sgdg\API_Client::get_directory_id( $root, $path[0] )->then(
+	return \Sgdg\API_Facade::get_directory_id( $root, $path[0] )->then(
 		static function( $next_dir_id ) use ( $path ) {
 			array_shift( $path );
 			return list_directories_in_path( $path, $next_dir_id );
