+++ src/php/frontend/gallery.php
@@ -70,7 +70,7 @@ function path_names( $path, $options ) {
 	return \Sgdg\Vendor\GuzzleHttp\Promise\Utils::all(
 		array_map(
 			static function( $segment ) use ( &$options ) {
-				return \Sgdg\API_Client::get_file_name( $segment )->then(
+				return \Sgdg\API_Facade::get_file_name( $segment )->then(
 					static function( $name ) use ( $segment, &$options ) {
 						$pos = false;
 						if ( '' !== $options->get( 'dir_prefix' ) ) {
