+++ src/php/skaut-google-drive-gallery.php
@@ -51,6 +51,7 @@ require_once __DIR__ . '/bundled/vendor-includes.php';
 
 require_once __DIR__ . '/class-options.php';
 require_once __DIR__ . '/class-api-client.php';
+require_once __DIR__ . '/class-api-facade.php';
 
 require_once __DIR__ . '/exceptions/class-exception.php';
 require_once __DIR__ . '/exceptions/class-api-exception.php';
